cloud_function/requirements.txt
requests>=2.31.0
google-auth>=2.29.0
